pyproject.toml
[project]
name = "crowdcounting-p2pnet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib==3.10.5",
    "numpy==2.3.2",
    "opencv-python==4.11.0.86",
    "pandas==2.3.1",
    "pillow==11.3.0",
    "scipy==1.16.1",
    "tensorboardx==2.6.4",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "tqdm==4.67.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]
